Publishing to Hash-Archive.org.
